Move plugins into plugins folder
