+++ packages/builder-vite/plugins/code-generator-plugin.ts
@@ -1,16 +1,16 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import { transformIframeHtml } from './transform-iframe-html';
-import { generateIframeScriptCode } from './codegen-iframe-script';
-import { generateModernIframeScriptCode } from './codegen-modern-iframe-script';
-import { generateImportFnScriptCode } from './codegen-importfn-script';
-import { generateVirtualStoryEntryCode, generatePreviewEntryCode } from './codegen-entries';
-import { generateAddonSetupCode } from './codegen-set-addon-channel';
+import { transformIframeHtml } from '../transform-iframe-html';
+import { generateIframeScriptCode } from '../codegen-iframe-script';
+import { generateModernIframeScriptCode } from '../codegen-modern-iframe-script';
+import { generateImportFnScriptCode } from '../codegen-importfn-script';
+import { generateVirtualStoryEntryCode, generatePreviewEntryCode } from '../codegen-entries';
+import { generateAddonSetupCode } from '../codegen-set-addon-channel';
 
 import type { Plugin } from 'vite';
-import type { ExtendedOptions } from './types';
+import type { ExtendedOptions } from '../types';
 
-import { virtualAddonSetupFile, virtualFileId, virtualPreviewFile, virtualStoriesFile } from './virtual-file-names';
+import { virtualAddonSetupFile, virtualFileId, virtualPreviewFile, virtualStoriesFile } from '../virtual-file-names';
 
 export function codeGeneratorPlugin(options: ExtendedOptions): Plugin {
   const iframePath = path.resolve(__dirname, '..', 'input', 'iframe.html');
